Optimized win condition check using WQUPC algorithm
Before, a numpy 2D array was used to store the game board.
To detect a win condition in this matrix, we had to consider
each possible row, column, diagonal and sub-diagonal of a given length.
Obviously, this is computationally complex problem (I won't even bother
estimating the complexity, but I guess it's superpolynomial)
and also not very efficient cause we store and consider empty elements.

The new approach is to store the game board in a hash table,
where the key is the cell's index and the value is the chip.
This is a sparse-matrix friendly solution.
Besides, it allows the following win condition detection optimization:
1. When a chip is placed on a grid, it forms a union with itself.
2. We check all the neighbors and unite the union of the newly occupied
   cell with the unions of the neighbor cells that store the same chip.
3. Union is done using Weighted Quick Union with Path Compression
   algorithms, that stores a forest in a hash table.
   Weighted modification of the algorithm also stores the size of each tree.
4. When we need to check for a win condition, we just get the tree with
   the max size and check whether this size is greater than the number
   of chips in a row we have to connect.

Overall, the new algorithm's complexity of detecting win condition is:
N + M lg* N, where N is the number of game states, M - the number of queries.